Make readers.py more easilly importable by moving away models
In recent versions of Django you can't import models until
late in the process, which is an issue because util.py does that
and is imported all over the place, including by readers.py.

Since readers.py is a kind of external API (mainly for FetchInProgress
we need to make sure it is easilly importable).